+++ android/app/src/main/kotlin/deckers/thibault/aves/channel/calls/DeviceHandler.kt
@@ -84,16 +84,12 @@ class DeviceHandler(private val context: Context) : MethodCallHandler {
             "script" to locale.script,
         )
 
+        // when called from a window-less service, locales from `context.resources`
+        // do not reflect the current system settings, so we use `Resources.getSystem()` instead
+        val list = Resources.getSystem().configuration.locales
         val locales = ArrayList<FieldMap>()
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-            // when called from a window-less service, locales from `context.resources`
-            // do not reflect the current system settings, so we use `Resources.getSystem()` instead
-            val list = Resources.getSystem().configuration.locales
-            for (i in 0..<list.size()) {
-                locales.add(toMap(list.get(i)))
-            }
-        } else {
-            locales.add(toMap(Locale.getDefault()))
+        for (i in 0..<list.size()) {
+            locales.add(toMap(list.get(i)))
         }
         result.success(locales)
     }